Finish split in modules
